App/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
chromadb>=0.6.0
sentence-transformers>=3.3.0
anthropic>=0.40.0
pdfplumber>=0.11.4
python-docx>=1.1.2
openpyxl>=3.1.5
watchdog>=6.0.0
aiofiles>=24.1.0
httpx>=0.27.0
